Refactor UI component organization
