+++ paradice/src/ui/pages/character_creation_page.cpp
@@ -1,5 +1,5 @@
 // ==========================================================================
-// Paradice Account Creation Page
+// Paradice Character Creation Page
 //
 // Copyright (C) 2020 Matthew Chaplain, All Rights Reserved.
 //
@@ -24,9 +24,7 @@
 //             OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
 //             SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 // ==========================================================================
-#include "paradice/ui/account_creation_page.hpp"
-
-#include "paradice/ui/detail/password_edit.hpp"
+#include "paradice/ui/pages/character_creation_page.hpp"
 
 #include <munin/aligned_layout.hpp>
 #include <munin/button.hpp>
@@ -51,32 +49,21 @@ namespace paradice::ui {
 // ==========================================================================
 // CONSTRUCTOR
 // ==========================================================================
-account_creation_page::account_creation_page()
+character_creation_page::character_creation_page()
 {
     auto name_edit = munin::make_edit();
-    auto password_edit = detail::make_password_edit();
-    auto repeat_password_edit = detail::make_password_edit();
 
     auto fields = munin::view(
-        munin::make_grid_layout({1, 3}),
+        munin::make_grid_layout({1, 1}),
         munin::make_framed_component(
-            munin::make_titled_frame("Name"), munin::make_viewport(name_edit)),
-        munin::make_framed_component(
-            munin::make_titled_frame("Password"),
-            munin::make_viewport(password_edit)),
-        munin::make_framed_component(
-            munin::make_titled_frame("Password (Repeat)"),
-            munin::make_viewport(repeat_password_edit)));
+            munin::make_titled_frame("Name"), munin::make_viewport(name_edit)));
 
     auto return_button = munin::make_button("Return");
     auto next_button = munin::make_button("Next");
 
     return_button->on_click.connect(on_return);
-    next_button->on_click.connect([=] {
-        on_next(
-            terminalpp::to_string(name_edit->get_text()),
-            terminalpp::to_string(password_edit->get_text()));
-    });
+    next_button->on_click.connect(
+        [=] { on_character_created(to_string(name_edit->get_text())); });
 
     auto buttons = munin::view(
         munin::make_compass_layout(),
@@ -88,17 +75,16 @@ account_creation_page::account_creation_page()
         munin::compass_layout::heading::east);
 
     set_layout(munin::make_grid_layout({1, 1}));
-    add_component(
-        munin::make_framed_component(
-            munin::make_titled_frame("Create New Account"),
-            munin::view(
-                munin::make_compass_layout(),
-                fields,
-                munin::compass_layout::heading::north,
-                munin::make_fill(' '),
-                munin::compass_layout::heading::centre,
-                buttons,
-                munin::compass_layout::heading::south)));
+    add_component(munin::make_framed_component(
+        munin::make_titled_frame("Create New Character"),
+        munin::view(
+            munin::make_compass_layout(),
+            fields,
+            munin::compass_layout::heading::north,
+            munin::make_fill(' '),
+            munin::compass_layout::heading::centre,
+            buttons,
+            munin::compass_layout::heading::south)));
 
     name_edit->set_focus();
 }
